// src/app/models/user.model.ts
export interface User {
    birth_date: Date;
    createdAt: Date;
    email: string;
    first_name: string;
    id: number;
    last_name: string;
    latitude: number;
    longitude: number;
    profile_image_url: string;
    show_birth_date: boolean;
    show_email: boolean;
    show_location: boolean;
    updatedAt: Date;
    address: string;
}